requirements.txt
mcp[cli]>=1.2.0
httpx>=0.24.0
requests==2.31.0
mcp-server>=0.1.0
python-dotenv>=0.19.0
unittest2>=1.1.0
pandas>=2.0.0
streamlit>=1.22.0
openpyxl>=3.1.2
tabulate>=0.9.0
reportlab>=4.0.4